requirements_dev.txt
colorama==0.3.1
github3.py==0.9.3
kyotocabinet==1.9
